SearchBar updates the query
