+++ src/components/SearchBar.vue
@@ -1,15 +1,16 @@
 <script setup lang="ts">
 import { ref } from 'vue'
+import { useRouter, useRoute } from 'vue-router'
 import { MdSearch } from '@vicons/ionicons4'
 import { NFlex, NInputGroup, NInput, NButton, NIcon } from 'naive-ui'
 
-const emit = defineEmits(['search-phrase'])
+const router = useRouter()
+const route = useRoute()
 
-const searchPhrase = ref('')
+const searchInput = ref('')
 
-function emitSearchPhrase() {
-  emit('search-phrase', searchPhrase.value)
-  searchPhrase.value = ''
+function pushSearchPhrase() {
+  router.push({ name: 'searchPage', query: { ...route.query, q: searchInput.value.trim() } })
 }
 </script>
 
@@ -21,10 +22,10 @@ function emitSearchPhrase() {
         placeholder="Search..."
         size="large"
         round
-        v-model:value="searchPhrase"
-        @keyup.enter="emitSearchPhrase"
+        v-model:value="searchInput"
+        @keyup.enter="pushSearchPhrase"
       />
-      <n-button circle size="large" @click="emitSearchPhrase" class="search-button">
+      <n-button circle size="large" @click="pushSearchPhrase" class="search-button">
         <n-icon>
           <md-search />
         </n-icon>
