+++ src/router/index.ts
@@ -5,12 +5,12 @@ const routes = [
   { path: '/', name: 'home', component: () => import('../views/HomeView.vue') },
   { path: '/upload', name: 'upload', component: () => import('../views/UploadView.vue') },
   { path: '/files/:id', name: 'filePage', component: () => import('../views/FileContentView.vue') },
+  { path: '/search', name: 'searchPage', component: () => import('../views/SearchPageView.vue') },
   {
     path: '/:pathMatch(.*)',
     name: 'notFound',
     component: () => import('../views/NotFoundView.vue'),
   },
-  { path: '/search', name: 'searchPage', component: () => import('../views/SearchPageView.vue') },
 ]
 
 const router = createRouter({
