+++ src/views/HomeView.vue
@@ -3,17 +3,10 @@ import { ref } from 'vue'
 import SearchBar from '../components/SearchBar.vue'
 import CategoryHolder from '../components/CategoryHolder.vue'
 import FileList from '../components/FileList.vue'
-
-const searchInput = ref('')
-
-function receiveEmit(receivedText: string) {
-  searchInput.value = receivedText
-}
 </script>
 
 <template>
-  <SearchBar @search-phrase="receiveEmit" />
-  <p>Input: {{ searchInput }}</p>
+  <SearchBar />
   <CategoryHolder />
   <FileList />
 </template>
